% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_cdnbcr_model.R
\name{cdnbcr_model}
\alias{cdnbcr_model}
\alias{dcdnbcr}
\alias{pcdnbcr}
\alias{rcdnbcr}
\alias{cure_rate}
\title{The Correlated Destructive Negative Binomial Cure Rate Model}
\usage{
dcdnbcr(x, theta, phi, p, alpha, mu, sigma, log.p = FALSE)

pcdnbcr(q, theta, phi, p, alpha, mu, sigma, lower.tail = TRUE, log.p = FALSE)

rcdnbcr(n, theta, phi, p, alpha, mu, sigma)

cure_rate(theta, phi, p, alpha)
}
\arguments{
\item{x}{Vector of positive event times or quantiles.}

\item{theta, phi}{Positive parameters associated with the expected initial number of competing
causes and their dispersion.}

\item{p, alpha}{Parameters controlling the probability of a cause remaining active. The correlation
among active causes is determined by \code{alpha}. Both \code{p} and
\code{alpha} are restricted to the closed unit interval \code{[0, 1]}.
When \code{alpha = 0}, the activations are assumed to be independent.}

\item{mu, sigma}{Strictly positive parameters of the Weibull survival distribution assumed for the
time-to-event of active competing causes.}

\item{log.p}{Logical; if \code{TRUE}, probabilities \code{p} are given as
\code{log(p)}.}

\item{q}{Vector of quantiles.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are
\code{P(X <= x)}, otherwise, \code{P(X > x)}.}

\item{n}{Number of random values to return.}
}
\value{
\code{dcdnbcr} returns the probability density function, \code{pcdnbcr} gives the
    distribution function, \code{cure_rate} returns the cure rate of the model, and \code{rcdnbcr}
    generates random observations.
}
\description{
Implements functions for the Correlated Destructive Negative Binomial Cure Rate (CDNBCR) model,
    including probability density, cumulative distribution, cure rate, and random generation.
}
\details{
The CDNBCR model describes the time until an event occurs while accounting for a cure fraction,
    assuming multiple initial competing causes with potential correlation. The model is
    destructive in the sense that some initial competing causes contributing to the occurrence
    of the event can be eliminated. Thus, the model is particularly useful in applications
    where interventions (e.g., treatments) may eliminate or inactivate some competing causes.

    We assume that the number of initial competing causes is described by a negative binomial
    random variable with mean \code{theta} and variance \code{theta * (1 + phi * theta)}. Each of
    the initial competing causes may remain contributing to the occurrence of the event with
    probability \code{p}. The model introduces dependence on the destruction or not of these
    initial causes, controlled by the parameter \code{alpha}. Finally, let \eqn{Y_j} be the
    survival time due to the \eqn{j}th remaining competing cause, \eqn{j = 1, \ldots, D},
    where \eqn{D} is a latent random variable. We assume that \eqn{Y_j} follows a Weibull distribution
    with mean \code{mu} and variance \code{mu^2 * (Gamma(2/sigma + 1) / Gamma(1/sigma + 1)^2 - 1)}.
    The observed uncensored survival time is given by \eqn{T = \textrm{min}(Y_1, \ldots, Y_D)}.
}
\examples{

# Parameters
theta <- 1.5
phi <- 1.2
p <- 0.6
alpha <- 0.3
mu <- 5
sigma <- 1.5

n <- 1000
tobs <- rcdnbcr(n, theta, phi, p, alpha, mu, sigma)

# Censoring
Cens <- 10
delta <- ifelse(tobs <= Cens, 1, 0)
tobs[delta == 0] <- Cens

op <- par()$mfrow
par(mfrow = c(1, 2))
# Histogram
hist(tobs, prob = TRUE, main = " ", xlab = expression(T[obs]))
curve(dcdnbcr(x, theta, phi, p, alpha, mu, sigma), add = TRUE, col = 2, lwd = 2)
points(Cens, mean(1 - delta), pch = 16, col = 2, lwd = 2)

# Survival function
plot(survival::survfit(survival::Surv(tobs, delta) ~ 1, se.fit = FALSE), xlab = expression(T[obs]),
     ylab = "Survival function")
curve(1 - pcdnbcr(x, theta, phi, p, alpha, mu, sigma), add = TRUE, col = 2, lwd = 2)

# Cure rate
abline(h = cure_rate(theta, phi, p, alpha), col = 4)
legend("bottomleft", c("Empirical", "Theoretical", "Cure rate"),
        col = c(1, 2, 4), bty = "n", lty = 1)
par(mfrow = op)
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>

Diego I. Gallardo <\email{diego.gallardo.mateluna@gmail.com}>
}
