% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_EM_algorithm.R
\name{control_EM}
\alias{control_EM}
\title{Control Parameters for the EM Algorithm}
\usage{
control_EM(method = "BFGS", maxit = 10000, start = NULL, prec = 5e-05, ...)
}
\arguments{
\item{method}{Optimization method for the "M" steps. The default method is \code{"BFGS"}.}

\item{maxit}{Maximum number of EM algorithm iterations. The default is \code{10.000} iterations.
If this limit is reached, a warning message is displayed.}

\item{start}{An optional vector of initial values for the algorithm. The expected order of the
parameters is \code{(beta1, phi, beta2, alpha, mu, sigma)}, where \code{beta1} and \code{beta2}
are vectors of coefficients associated with the regression structures of
\code{theta} and \code{p}, respectively.}

\item{prec}{Numeric tolerance for convergence in the EM iterations.}

\item{...}{Additional arguments passed to \code{\link[stats]{optim}}.}
}
\value{
A list containing the specified control parameters.
}
\description{
A list of control parameters for the Expectation-Maximization (EM) algorithm used in the
    estimation of the Correlated Destructive Negative Binomial Cure Rate Model.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
