% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_EM_algorithm.R
\name{EM}
\alias{EM}
\title{EM Algorithm for Model Estimation}
\usage{
EM(
  time,
  delta,
  Z1 = NULL,
  Z2 = NULL,
  theta.link = "log",
  p.link = "logit",
  alpha = TRUE,
  control = control_EM(...),
  ...
)
}
\arguments{
\item{time}{A vector of positive response variables (the follow up time).}

\item{delta}{Status indicator where \code{0} represents right censoring, and \code{1} indicates an
event occurrence at the observed time.}

\item{Z1, Z2}{Design matrices for the regression structures of \code{theta} and \code{p}, respectively.}

\item{theta.link, p.link}{Link functions for the regression models of \code{theta} and \code{p}.
The default values are \code{"log"} for \code{theta.link} and \code{"logit"} for \code{p.link}.
Other options are available via \code{\link[stats]{make.link}}. The choice of link functions
should consider the possible values of \code{theta} (positive values) and \code{p} (unit interval [0,1]).}

\item{alpha}{If \code{TRUE} (default), \code{alpha} is included in the estimation process to
model correlation between the initial competing causes. If \code{FALSE}, \code{alpha} is
fixed at \code{0}, reducing the model to the uncorrelated case.}

\item{control}{A list of control parameters for the EM algorithm, specified using \code{\link{control_EM}}.}

\item{...}{Additional arguments passed to \code{\link{control_EM}}.}
}
\value{
\code{EM} returns a list with the following components:
\describe{
  \item{estimates}{ A vector containing the estimated model parameters.}
  \item{par}{A named list of parameter estimates, that is, a list containing the elements
      \code{beta1}, \code{phi}, \code{beta2}, \code{alpha}, \code{mu}, and \code{sigma}.}
  \item{iterations}{ The number of EM iterations performed.}
  \item{convergence}{An integer indicating the convergence status, where \code{0} indicates
      successful convergence and \code{1} indicates that the maximum number of iterations has been reached.}
  \item{vcov}{The estimated covariance matrix of the estimated parameters.}
  \item{latent}{Latent variables obtained during the EM algorithm.}
 }
}
\description{
An EM algorithm for the maximum likelihood estimation of the parameters of the Correlated
    Destructive Negative Binomial Cure Rate Model.
}
\author{
Diego I. Gallardo \email{diego.gallardo.mateluna@gmail.com}

Rodrigo M. R. de Medeiros \email{rodrigo.matheus@ufrn.br}
}
