% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_methods.R
\name{summary.cdnbcr}
\alias{summary.cdnbcr}
\alias{print.summary.cdnbcr}
\title{Summarizing a CDNBCR Fit}
\usage{
\method{summary}{cdnbcr}(object, ...)

\method{print}{summary.cdnbcr}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class \code{"cdnbcr"}, a result of a call to \code{\link{cdnbcr}}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class \code{"summary.cdnbcr"}, a result of a call to \code{summary.cdnbcr}.}

\item{digits}{A non-null value for digits specifies the minimum number of significant digits to
be printed in values.}
}
\value{
The function \code{summary.cdnbcr} returns an object of class \code{"summary.cdnbcr"},
  a list with the following components:
\describe{
  \item{call}{The original function call.}
  \item{theta}{Summary of the regression coefficients for the \code{theta} submodel.}
  \item{p}{Summary of the regression coefficients for the \code{p} submodel.}
  \item{par}{Estimates and standard errors for the additional model parameters: \code{phi}
    (dispersion of the initial number of competing causes), \code{alpha} (dependence parameter,
    when estimated), and the baseline parameters \code{mu} and \code{sigma} (reparameterized Weibull).}
  \item{links}{Named list with the link functions used in the \code{theta} and \code{p}
    regression submodels.}
  \item{residuals}{A \code{"\link[survival]{Surv}"} object with the Cox-Snell residuals
    (Cox and Snell, 1968). If the model is well fitted to the data, the Cox-Snell residuals are
     expected to be distributed as a censored random sample from the exponential distribution
      with mean 1.}
  \item{iterations}{Number of EM iterations.}
  \item{logLik}{Log-likelihood of the fitted model.}
  \item{AIC, BIC}{Akaike and Bayesian information criteria.}
}
}
\description{
\code{summary} method for class \code{"cdnbcr"}.
}
\examples{
\donttest{## Loading survival package
library(survival)

## Dataset: e1690 (see ?e1690)
head(e1690)

## Correlated destructive fit
fit <- cdnbcr(formula = Surv(time, status) ~ nodeII + nodeIII + nodeIV - 1 |
                sex + trt + thickness + age, data = e1690)
out <- summary(fit)
out

### Summary table for the regression coefficients
out$theta
out$p

### Summary table for the additional parameters
out$par

### Cox-Snell residuals
out$residuals
class(out$residuals)
plot(out$residuals)
curve(pexp(x, 1, lower.tail = FALSE), add = TRUE, col = "blue")

## Uncorrelated destructive fit
fit0 <- cdnbcr(formula = Surv(time, status) ~ nodeII + nodeIII + nodeIV - 1 |
                 sex + trt + thickness + age, data = e1690, alpha = FALSE)
summary(fit0)}
}
\references{
Cox, D. R., and Snell, E. J. (1968). A general definition of residuals.
    \emph{Journal of the Royal Statistical Society B}, \bold{30}, 248--265.

De Medeiros, R. M. R., Bourguignon, M., Gómez, Y. M., and Gallardo, D. I. (2026).
    A correlated approach to cancer cell counting in cure rate models.

Rodrigues, J., De Castro, M., Balakrishnan, N., and Cancho, V. G. (2011).
    Destructive weighted Poisson cure rate models. \emph{Lifetime Data Analysis}, \bold{17}, 333--346
}
\author{
Diego I. Gallardo \email{diego.gallardo.mateluna@gmail.com}

Rodrigo M. R. de Medeiros \email{rodrigo.matheus@ufrn.br}
}
