% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_methods.R
\name{cdnbcr-methods}
\alias{cdnbcr-methods}
\alias{model.frame.cdnbcr}
\alias{model.matrix.cdnbcr}
\alias{print.cdnbcr}
\alias{summary.cdnbcr}
\alias{print.summary.cdnbcr}
\alias{coef.cdnbcr}
\alias{logLik.cdnbcr}
\alias{AIC.cdnbcr}
\alias{vcov.cdnbcr}
\title{Methods for "cdnbcr" objects}
\usage{
\method{model.frame}{cdnbcr}(formula, ...)

\method{model.matrix}{cdnbcr}(object, submodel = c("theta", "p"), ...)

\method{print}{cdnbcr}(x, digits = getOption("digits"), ...)

\method{summary}{cdnbcr}(object, ...)

\method{print}{summary.cdnbcr}(x, digits = getOption("digits"), ...)

\method{coef}{cdnbcr}(object, submodel = c("full", "theta", "p"), ...)

\method{logLik}{cdnbcr}(object, ...)

\method{AIC}{cdnbcr}(object, ..., k = 2)

\method{vcov}{cdnbcr}(object, submodel = c("full", "theta", "p"), ...)
}
\arguments{
\item{formula}{A model \code{\link[Formula]{Formula}} or \code{\link[stats]{terms}} object or an
\code{"cdnbcr"} object.}

\item{...}{Additional arguments passed to or from other methods.}

\item{submodel}{Specifies which part of the model should be used. Options:
\code{"full"} (default) for all parameters, \code{"theta"} for the regression submodel of
\code{theta}, and \code{"p"} for the regression submodel of\code{p}.}

\item{x, object}{An object of class \code{"cdnbcr"}.}

\item{digits}{Minimal number of significant digits, see \code{\link[base]{print.default}}.}

\item{k}{Numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC. See \code{\link[stats]{AIC}}.}
}
\description{
Methods for "cdnbcr" objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
