% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_fit_function.R
\name{cdnbcr}
\alias{cdnbcr}
\title{Correlated Destructive Negative Binomial Cure Rate Model}
\usage{
cdnbcr(
  formula,
  data,
  subset,
  na.action,
  theta.link = "log",
  p.link = "logit",
  alpha = TRUE,
  control = control_EM(...),
  y = FALSE,
  x = FALSE,
  ...
)
}
\arguments{
\item{formula}{A symbolic description of the model. It must include a \code{\link[survival]{Surv}}
object as the response on the left-hand side of the \code{~} operator. The right-hand side
of the formula can contain covariates for two sub-models: one for \code{theta} (expected number
of initial competing causes) and one for \code{p} (probability of activation of an initial
competing cause). Covariates are separated by the \code{|} symbol. See details below.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{subset}{An optional vector specifying a subset of observations to be used in the model.}

\item{na.action}{A function to handle missing values, see \code{\link[stats]{model.frame}}.}

\item{theta.link, p.link}{Link functions for the regression submodels of \code{theta} and \code{p}.
Defaults are \code{"log"} for \code{theta.link} and \code{"logit"} for \code{p.link}.
Other options are available via \code{\link[stats]{make.link}}.}

\item{alpha}{Logical; if \code{TRUE} (default), \code{alpha} is estimated to model correlation
among initial competing causes. If \code{FALSE}, \code{alpha} is fixed at 0, reducing the
model to the uncorrelated case.}

\item{control}{A list of control parameters for the EM algorithm, specified using \code{\link{control_EM}}.}

\item{y}{Logical; if \code{TRUE}, the response vector is returned.}

\item{x}{Logical; if \code{TRUE}, the model matrices are returned.}

\item{...}{Additional arguments passed to \code{\link{control_EM}}.}
}
\value{
The function \code{cdnbcr} returns an object of class \code{"cdnbcr"}, which consists of a list
    with the following components:
\describe{
  \item{coefficients}{A list containing the vectors of the regression coefficients \code{beta1}
       and \code{beta2}, associated with the regression structures of \code{theta} and \code{p},
        respectively.
  }
  \item{phi, alpha, mu, sigma}{Estimates of the additional parameters of the model.}
  \item{fitted}{Fitted values for \code{theta} and \code{p}}
  \item{residuals}{Vector of Cox-Snell residuals.}
  \item{vcov}{The estimated covariance matrix of the estimated parameters.}
  \item{logLik}{Log-likelihood of the fitted model.}
  \item{nobs}{Number of observations.}
  \item{df.null}{Residual degrees of freedom in the null model.}
  \item{df.residual}{Residual degrees of freedom in the fitted model.}
  \item{convergence}{Integer code indicating convergence status (0 = successful, 1 = max iterations reached).}
  \item{inits}{Initial values used in the EM algorithm.}
  \item{control}{List of control parameters used in the EM algorithm.}
  \item{EM_iterations}{Number of EM iterations performed.}
  \item{call}{Function call.}
  \item{formula}{Model formula used.}
  \item{terms}{List of terms objects for the two regression frameworks.}
  \item{y}{Response vector, if \code{y = TRUE}.}
  \item{x}{Model matrices, if \code{x = TRUE}.}
 }
}
\description{
Implements an Expectation-Maximization (EM) algorithm for maximum likelihood estimation in the
   Correlated Destructive Negative Binomial Cure Rate Model.
}
\details{
The \code{formula} argument follows the syntax of the \code{Formula} package
    (Zeileis and Croissant, 2010) that allows specifying regression structures for more than one
     parameter in the model. For example, suppose the data are formed as follows:

 \tabular{cccccccccccc}{
 \bold{time}   \tab \tab \bold{status} \tab \tab \bold{x1} \tab \tab \bold{x2} \tab \tab \bold{z1} \tab \tab \bold{z2} \cr
 time_1  \tab \tab status_1 \tab \tab x1_1 \tab \tab x2_1 \tab \tab z1_1 \tab \tab z2_1 \cr
 time_2  \tab \tab status_2 \tab \tab x1_2 \tab \tab x2_2 \tab \tab z1_2 \tab \tab z2_2 \cr
 ...  \tab \tab ... \tab \tab... \tab \tab ... \tab \tab ... \tab \tab ...  \cr
 time_n  \tab \tab status_n \tab \tab x1_n \tab \tab x2_n \tab \tab z1_n  \tab \tab z2_n \cr
 }
where, for each element of the sample, \code{time} is the follow-up time; \code{status}
indicates whether the observed time refers to the event of interest (\code{status = 1}) or
whether it was right-censored (\code{status = 0}); and \code{x1, x2, z1, z2} are
explanatory variables. Then, \code{formula = Surv(time, status) ~ x1 + x2 | z1 + z2}
specifies a CDNBCR model with two regression structures, one for \code{theta} in terms of the
explanatory variables \code{x1} and \code{x2} and another for \code{p} in terms of the explanatory
variables \code{z1} and \code{z2}.
}
\examples{
 \dontrun{
  ## Loading survival package
  library(survival)

  ## Data (see ?e1690)
  head(e1690)

  ## Correlated destructive fit
  fit <- cdnbcr(formula = Surv(time, status) ~ nodeII + nodeIII + nodeIV - 1 |
                sex + trt + thickness + age, data = e1690)

  summary(fit)

  ## Cox-Snell residuals
  par(mfrow = c(1, 2))
  plot(fit, ask = FALSE)
  par(mfrow = c(1, 1))

  ## Latent variables
  plot(fit$latent$M, ylab = "Initial competing causes", pch = 16, cex = 0.8)
  plot(fit$latent$D, ylab = "Remaining competing causes", pch = 16, cex = 0.8)

  ## Uncorrelated destructive fit (alpha = 0)
  fit0 <- cdnbcr(formula = Surv(time, status) ~ nodeII + nodeIII + nodeIV - 1 |
                 sex + trt + thickness + age, alpha = FALSE, data = e1690)

  summary(fit0)
 }
}
\author{
Diego I. Gallardo \email{diego.gallardo.mateluna@gmail.com}

Rodrigo M. R. de Medeiros \email{rodrigo.matheus@ufrn.br}
}
