% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_EM_algorithm.R
\name{control_EM}
\alias{control_EM}
\title{Auxiliary for Controlling a CDNBCR Fit via EM Algorithm}
\usage{
control_EM(method = "BFGS", maxit = 10000, start = NULL, prec = 5e-05, ...)
}
\arguments{
\item{method}{Optimization method for the "M" steps, specifying the \code{method} argument
passed to \code{\link[stats]{optim}}.}

\item{maxit}{Maximum number of EM algorithm iterations. The default is \code{10.000} iterations.
If this limit is reached, a warning message is displayed.}

\item{start}{An optional vector of initial values for the algorithm. The expected order of the
parameters is \code{(beta1, phi, beta2, alpha, mu, sigma)}, where \code{beta1} and \code{beta2}
are vectors of coefficients associated with the regression structures of
\code{theta} and \code{p}, respectively.}

\item{prec}{Numeric tolerance for convergence in the EM iterations.}

\item{...}{Additional arguments passed to \code{\link[stats]{optim}}.}
}
\value{
A list with components named as the arguments.
}
\description{
Optimization parameters that control the fitting of the correlated destructive negative
    binomial cure rate model (CDNBCR) via Expectation-Maximization (EM) algorithm.
}
\examples{
\donttest{## Loading the survival package
library(survival)

## Dataset:  (see ?e1690)
head(e1690)

## Correlated destructive fit (default fit)
fit1 <- cdnbcr(Surv(time, status) ~ nodeII + nodeIII + nodeIV - 1 |
                trt + thickness, data = e1690)

## Changing the initial values
beta1 <- c(1, 1.5, 2)
phi <- 2
beta2 <- c(0.5, -0.5, 0.5)
alpha <- 0.5
mu <- 3
sigma <- 2

fit2 <- cdnbcr(Surv(time, status) ~ nodeII + nodeIII + nodeIV - 1 |
                trt + thickness, data = e1690,
               start = c(beta1, phi, beta2, alpha, mu, sigma))

## Compare the fits:
fit1
fit2}
}
\seealso{
\code{\link{cdnbcr}} for parameter estimation in the CDNBCR model via EM algorithm.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
