% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_methods.R
\name{plot.cdnbcr}
\alias{plot.cdnbcr}
\title{Diagnostic Plots for the Correlated Destructive Negative Binomial Cure Rate
    Model}
\usage{
\method{plot}{cdnbcr}(
  x,
  which = 1:2,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{x}{An object of class \code{cdnbcr}.}

\item{which}{numeric; if a subset of the plots is required, specify a subset
of the numbers \code{1:2}.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Produces diagnostic plots to assess the goodness-of-fit of a Correlated Destructive Negative
    Binomial Cure Rate model fit based on the Cox-Snell residuals. Available plots include the
    Kaplan-Meier estimate and the cumulative hazard plot.
}
\details{
The function produces two diagnostic plots for assessing model fit:
(1) The Kaplan-Meier estimate for Cox-Snell residuals, compared to the expected
    survival function of an exponential distribution with mean 1; and
(2) The cumulative hazard function against the Cox-Snell residuals, which
    should align approximately with the identity line if the model is well-fitted.
}
