% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_methods.R
\name{predict.cdnbcr}
\alias{predict.cdnbcr}
\title{Predict Method for the Correlated Destructive Negative Binomial Cure
    Rate Regression}
\usage{
\method{predict}{cdnbcr}(
  object,
  newdata = NULL,
  type = c("cure", "survival"),
  na.action = stats::na.pass,
  ...
)
}
\arguments{
\item{object}{A \code{"cdnbcr"} object.}

\item{newdata}{Optionally, a data frame in which to look for variables
with which to predict. If omitted, the fitted linear predictors are
used.}

\item{type}{The type of prediction required. The default is the
fitted cure rate (\code{"cure"}). The alternative \code{"survival"}
provides the fitted survival curve.}

\item{na.action}{function determining what should be done with missing
values in \code{newdata}. The default is to predict \code{NA}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A vector of predictions.
}
\description{
Obtains predictions from a fitted \code{"cdnbcr"} object.
}
